Documentation and bug fixes
