+++ client/README.md
@@ -1,0 +1,43 @@
+# Client - Mercado Libre Challenge 
+
+This folder containst all related with the client
+
+## File Structure
+```
+.
+├── package.json
+├── README.md
+├── public
+│   ├── index.html
+│   ├── manifest.json
+├── src
+│   ├── App.css
+│   ├── App.js
+│   ├── index.css
+│   ├── index.js
+│   ├── api
+│   │   ├── config.js
+│   │   ├── requestSender.js
+│   ├── assets
+│   │   ├── png
+│   │   │   ├── ic_not_found.png
+│   │   │   ├── ic_Search.png
+│   │   │   ├── ic_shipping.png
+│   │   │   ├── Logo_ML.png
+│   ├── components
+│   │   ├── utils
+│   │   │   ├── CategoryPath.js
+│   │   │   ├── PriceFormatter.js
+│   │   │   ├── Translate.js
+│   ├── config
+│   │   ├── routes
+│   ├── pages
+│   │   ├── ItemDetail.js
+│   │   ├── ItemList.js
+│   │   ├── SearchBox.js
+│   ├── scss
+│   │   ├── _variable.scss
+│   │   ├── ItemDetail.scss
+│   │   ├── ItemList.scss
+│   │   ├── SearchBox.scss
+```
