+++ client/src/pages/ItemDetail.js
@@ -2,7 +2,8 @@ import React from "react";
 import { Layout, Button, Carousel } from "antd";
 import { Route, Redirect } from "react-router-dom";
 import PriceFormatter from "../components/utils/PriceFormatter";
-import CategoryPathRender from "../components/utils/CategoryPath"
+import CategoryPathRender from "../components/utils/CategoryPath";
+import {Condition, SoldQuantity} from "../components/utils/Translate";
 
 import SearchBox from "../pages/SearchBox";
 
@@ -11,8 +12,8 @@ import '../scss/ItemDetail.scss';
 /**
  * The ItemDetail component
  * 
- * This component renders the item detail layout. When the user sends select an item 
- * through the click on it, the front-end sends the petition to the back-end and 
+ * This component renders the item detail layout. When the user sends an item 
+ * by clicking on it, the front-end sends the request to the back-end and 
  * renders the detail.
  * 
  * @param {object} props the props sending from the latest component.
@@ -70,40 +71,6 @@ export default function ItemDetail(props){
             </Content>
         </Layout>
     );
-}
-
-/**
- * 
- * Condition component
- * 
- * This component map the condition of the item in order to show the item condition in Spanish
- * 
- * @param {boolean} condition The condition variable is assigned by destructuring and it is
- *      a flag that indicates if the item is new or used in the condition param 
- */
-function Condition({condition}){
-    const conditionLabels =  {
-        "used": "Usado",
-        "new": "Nuevo"
-    }
-    return conditionLabels[condition];
-}
-
-/**
- * SoldQuantity component
- * 
- * This component returns the appropriate word (in singular or plural) according to the sold quantity
- * 
- * @param {Integer} soldQuantity The soldQuantity variable is assigned by destructuring and it is
- *      a flag that indicates the sold units of the item. 
- */
-
-function SoldQuantity({soldQuantity}){
-    if(soldQuantity !== 1){
-        return <span> {soldQuantity} vendidos</span>;
-    }else{
-        return <span> {soldQuantity} vendido</span>;
-    }
 }
 
 /**
