+++ server/README.md
@@ -1,0 +1,22 @@
+# Server - Mercado Libre Challenge 
+
+This folder containst all related with the server
+
+## File Structure
+```
+.
+├── package.json
+├── README.md
+├── app.js
+├── config.js
+├── controllers
+│   ├── ItemControler.js
+├── routes
+│   ├── routes.js
+├── services
+│   ├── RequestService.js
+│   ├── UtilService.js
+├── test
+│   ├── ItemControllerTest.js
+
+```
