+++ server/services/UtilService.js
@@ -1,5 +1,14 @@
 const { URL } = require('url');
 
+/**
+ * 
+ * buildUrl function
+ * 
+ * This function builds the targetURL to send to the back-end.
+ * 
+ * @param {string} path the path to send the request
+ * @param {object} queryParams the object with the query params
+ */
 function buildURL(path, queryParams){
     const baseUrl = new URL(path, 'https://api.mercadolibre.com');
     if(queryParams){
